Rename file to match exported type
