+++ src/model/uniprotkb/sections/PathologyAndBiotechConverter.ts
@@ -6,7 +6,7 @@ import {
   Keyword,
   KeywordUIModel,
 } from '../../utils/KeywordsUtil';
-import KeywordCategories from '../../types/KeywordTypes';
+import KeywordCategories from '../../types/KeywordCategories';
 import FeatureTypes from '../../types/FeatureTypes';
 
 type PathologyAndBiotechAPIModel = {
